+++ sembast/lib/utils/key_utils.dart
@@ -1,0 +1,3 @@
+library sembast.utils.key_utils;
+
+export 'package:sembast/src/key_utils.dart' show generateStringKey;